new diagrams model

+++ Core/Geometries/Range.cs
@@ -4,11 +4,11 @@
 
     public struct Range
     {
-        private double min;
-        private double max;
+        private int min;
+        private int max;
 
 
-        public Range(double x, double y)
+        public Range(int x, int y)
         {
             if (x > y)
             {
@@ -22,17 +22,17 @@
             }
         }
 
-        public bool Contains(double value)
+        public bool Contains(int value)
         {
             return Range.Contains(min, max, value);
         }
 
-        public static bool Contains(double x, double y, double value)
+        public static bool Contains(int x, int y, int value)
         {
             return value >= Math.Min(x, y) && value <= Math.Max(x, y);
         }
 
-        public static bool Contains(double x,double y,double value,double delta)
+        public static bool Contains(int x,int y,int value,int delta)
         {
             return value >= Math.Min(x, y) - Math.Abs(delta) && 
                 value <= Math.Max(x, y) + Math.Abs(delta);
